Uploaded Session ID proof of concept html & phhp
The file session.html is an experimental page with 2 PHP scripts, page1.phph and page2.php. With instructions on the page, one can see how PHP sessions are created, restored, and deleted. Only a proof of concept - more work needs to be done, not to mention implementing this as a stand-alone PHP script that will work on our own website. Note: the phph scripts may not work if not run from an actual website server.