Added View Personal Information & Check Valid Session ID
The webpage and script for viewing one's personal information are here. These files can be accessed from John's pages which I had to add a link to, and which I'll upload in a moment. Also included is a PHP code fro checking one's Session ID for validity. Currently, this code can be copy-pasted in other codes, but most likely it'll be slightly changed to be 'called' instead of copy-pasted.