Added & slightly modified John's login and logout scripts
I had to add a link to John's 'mainmenu.php' in order to access the View Personal Information I designed. Another slight change in the PHP code (but bigger in the database) is the addition of an attribute to the Session ID table, one containing the Username. This was essential for viewing one's personal info.