Implemented the Create Order task
The files createorder and addtoorder were added for this purpose. One generates the order with an incremental orderid, the other generates entires in the OrderContain table of the database. That table stores the items associated with each order. Mainmene.php was slightly modified to present a link to createorder.php.